--- samples/examples/Reggression_BikeSharingDemands/BikeSharingDemand.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27428.2037
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BikeSharingDemand", "BikeSharingDemand\BikeSharingDemand.csproj", "{DD5EBF32-C310-405A-BD8D-507A35F0C204}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{DD5EBF32-C310-405A-BD8D-507A35F0C204}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DD5EBF32-C310-405A-BD8D-507A35F0C204}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DD5EBF32-C310-405A-BD8D-507A35F0C204}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DD5EBF32-C310-405A-BD8D-507A35F0C204}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {98369941-33DD-450C-A410-B9A91C8CDE91}
-	EndGlobalSection
-EndGlobal
