--- samples/examples/Reggression_BikeSharingDemands/BikeSharingDemand/BikeSharingDemandData/BikeSharingDemandPrediction.cs
@@ -1,10 +1,0 @@
-﻿using Microsoft.ML.Runtime.Api;
-
-namespace BikeSharingDemand.BikeSharingDemandData
-{
-    public class BikeSharingDemandPrediction
-    {
-        [ColumnName("Score")]
-        public float PredictedCount;
-    }
-}
