+++ samples/examples/Regression_BikeSharingDemand/BikeSharingDemand/BikeSharingDemandData/BikeSharingDemandPrediction.cs
@@ -1,0 +1,10 @@
+﻿using Microsoft.ML.Runtime.Api;
+
+namespace BikeSharingDemand.BikeSharingDemandData
+{
+    public class BikeSharingDemandPrediction
+    {
+        [ColumnName("Score")]
+        public float PredictedCount;
+    }
+}
